Updated intermediate_net_in_tensorflow
Had to change the data loading because tf.contrib is not stable. TensorFlow has not put a decent solution in place for quickly loading the MNIST data without tf.contrib so I just defaulted to using the keras loading. I'm not sure it matters where we get the data from TBH.